requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
idna==3.10
load-dotenv==0.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
Pygments==2.19.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.4
schedule==1.2.2
tweepy==4.14.0
urllib3==2.3.0
psycopg2-binary
sqlalchemy
alembic
openai